+++ src/functions/function_definition.ts
@@ -1,0 +1,123 @@
+import {
+  ManifestFunctionSchema,
+  ManifestFunctionType,
+} from "../manifest/manifest_schema.ts";
+import {
+  ParameterSetDefinition,
+  PossibleParameterKeys,
+} from "../parameters/types.ts";
+import {
+  FunctionDefinitionArgs,
+  ISlackFunctionDefinition,
+  SlackFunctionDefinitionArgs,
+} from "./types.ts";
+import { SlackManifest } from "../manifest/mod.ts";
+
+abstract class BaseFunctionDefinition<
+  InputParameters extends ParameterSetDefinition,
+  OutputParameters extends ParameterSetDefinition,
+  RequiredInput extends PossibleParameterKeys<InputParameters>,
+  RequiredOutput extends PossibleParameterKeys<OutputParameters>,
+> implements
+  ISlackFunctionDefinition<
+    InputParameters,
+    OutputParameters,
+    RequiredInput,
+    RequiredOutput
+  > {
+  public id: string;
+  abstract type: ManifestFunctionType;
+
+  constructor(
+    public definition: FunctionDefinitionArgs<
+      InputParameters,
+      OutputParameters,
+      RequiredInput,
+      RequiredOutput
+    >,
+  ) {
+    this.id = definition.callback_id;
+    this.definition = definition;
+  }
+
+  registerParameterTypes(manifest: SlackManifest) {
+    const { input_parameters: inputParams, output_parameters: outputParams } =
+      this.definition;
+    manifest.registerTypes(inputParams?.properties ?? {});
+    manifest.registerTypes(outputParams?.properties ?? {});
+  }
+
+  abstract export(): ManifestFunctionSchema;
+}
+
+export class SlackFunctionDefinition<
+  InputParameters extends ParameterSetDefinition,
+  OutputParameters extends ParameterSetDefinition,
+  RequiredInput extends PossibleParameterKeys<InputParameters>,
+  RequiredOutput extends PossibleParameterKeys<OutputParameters>,
+> extends BaseFunctionDefinition<
+  InputParameters,
+  OutputParameters,
+  RequiredInput,
+  RequiredOutput
+> {
+  type: ManifestFunctionType = "app";
+  constructor(
+    public definition: SlackFunctionDefinitionArgs<
+      InputParameters,
+      OutputParameters,
+      RequiredInput,
+      RequiredOutput
+    >,
+  ) {
+    super(definition);
+  }
+
+  export(): ManifestFunctionSchema {
+    return {
+      title: this.definition.title,
+      description: this.definition.description,
+      source_file: this.definition.source_file,
+      input_parameters: this.definition.input_parameters ??
+        { properties: {}, required: [] },
+      output_parameters: this.definition.output_parameters ??
+        { properties: {}, required: [] },
+    };
+  }
+}
+
+export class ConnectorFunctionDefinition<
+  InputParameters extends ParameterSetDefinition,
+  OutputParameters extends ParameterSetDefinition,
+  RequiredInput extends PossibleParameterKeys<InputParameters>,
+  RequiredOutput extends PossibleParameterKeys<OutputParameters>,
+> extends BaseFunctionDefinition<
+  InputParameters,
+  OutputParameters,
+  RequiredInput,
+  RequiredOutput
+> {
+  type: ManifestFunctionType = "API";
+  constructor(
+    public definition: FunctionDefinitionArgs<
+      InputParameters,
+      OutputParameters,
+      RequiredInput,
+      RequiredOutput
+    >,
+  ) {
+    super(definition);
+  }
+
+  export(): ManifestFunctionSchema {
+    return {
+      title: this.definition.title,
+      description: this.definition.description,
+      type: this.type,
+      input_parameters: this.definition.input_parameters ??
+        { properties: {}, required: [] },
+      output_parameters: this.definition.output_parameters ??
+        { properties: {}, required: [] },
+    };
+  }
+}
